% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oFDAinfo-package.R
\docType{package}
\name{oFDAinfo-package}
\alias{oFDAinfo}
\alias{oFDAinfo-package}
\title{oFDAinfo: Pulls Data From The FDA openAPI}
\description{
oFDAinfo pulls data from the FDA's openAPI. The package allows users to specify drugs of interest, and the data available will be collated and output as a dataframe (and a saved to a file if desired). As of now the functions only work on NDC type queries within the FDA.
}
\author{
\strong{Maintainer}: Jonah Gorodensky \email{jonahgoro@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
